+++ docs/html/search/all_6.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['write_11',['write',['../classace__wire_1_1SimpleWireFastInterface.html#a6fce64dbf641b21c3eadcb87017e02d5',1,'ace_wire::SimpleWireFastInterface::write()'],['../classace__wire_1_1SimpleWireInterface.html#a6c5a9cb8042d43bc86069fa950128687',1,'ace_wire::SimpleWireInterface::write()'],['../classace__wire_1_1TwoWireInterface.html#a4e28da3a3e6ae9a799ce1573da7ce6f5',1,'ace_wire::TwoWireInterface::write()']]]
+  ['write_10',['write',['../classace__wire_1_1SimpleWireFastInterface.html#a6fce64dbf641b21c3eadcb87017e02d5',1,'ace_wire::SimpleWireFastInterface::write()'],['../classace__wire_1_1SimpleWireInterface.html#a6c5a9cb8042d43bc86069fa950128687',1,'ace_wire::SimpleWireInterface::write()'],['../classace__wire_1_1TwoWireInterface.html#a4263936c9bff1c092855ea49bcd12e85',1,'ace_wire::TwoWireInterface::write()']]]
 ];